Refactoring Recipe class
Getting rid of the arrays in the variables. Fixing some mistakes. Adding room db to the gradle.

+++ app/src/main/java/fi/group6/dessertrecipeapp/classes/Ingredient.java
@@ -19,7 +19,7 @@ public class Ingredient {
      * @param measure
      * Measure for the amount ( kg, l, etc... )
      */
-    public Ingredient ( String name, int amount, String measure ) {
+    public Ingredient ( String name, double amount, String measure ) {
         this.name = name;
         this.amount = amount;
         this.measure = measure;
